requirements.txt
alembic==1.7.7
apturl==0.5.2
asgiref==3.5.2
autopep8==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
certifi==2022.5.18.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
cloudinary==1.29.0
colorama==0.4.4
command-not-found==0.3
config==0.5.1
confusable-homoglyphs==3.2.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dj-database-url==0.5.0
Django==4.0.2
django-bootstrap3==21.2
django-bootstrap5==21.3
django-braces==1.15.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-jquery==3.1.0
django-model-utils==4.2.0
django-registration==3.3
django-star-ratings==0.9.2
django-voting==1.0
djangorestframework==3.13.1
dnspython==2.2.1
dominate==2.6.0
duplicity==0.8.21
email-validator==1.2.1
fasteners==0.14.1
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Reuploaded==1.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
future==0.18.2
gpg===1.16.0-unknown
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==1.1.0
jeepney==0.7.1
Jinja2==2.11.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.0
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.1.1
protobuf==3.12.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-decouple==3.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
swapper==1.3.0
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
visitor==0.1.3
wadllib==1.3.6
Werkzeug==0.16.0
whitenoise==6.2.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==3.8.0
